Feat: Add job application listing page with applicant score calculation using OpenAI PDF analysis, and restrict access permissions to job_seeker users only.

+++ job-backoffice/database/migrations/2026_03_14_140910_create_job_applications_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('job_applications', function (Blueprint $table) {
             $table->uuid('id')->primary();
             $table->enum('status', ['pending', 'accepted', 'rejected'])->default('pending');
-            $table->float('aiGeneratedScore', 2)->default(0);
+            $table->decimal('aiGeneratedScore', 3, 1)->default(0);
             $table->mediumText('aiGeneratedFeedback')->nullable();
             $table->softDeletes();
             $table->timestamps();
